[lldb] Extract reusable POSIX register contexts
