+++ lldb/source/Plugins/Process/Common/RegisterContext/POSIX/RegisterContextPOSIX_powerpc.h
@@ -1,4 +1,4 @@
-//===-- RegisterContextPOSIX_ppc64le.h --------------------------*- C++ -*-===//
+//===-- RegisterContextPOSIX_powerpc.h --------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,20 +6,23 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_PPC64LE_H
-#define LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_PPC64LE_H
+#ifndef LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_POWERPC_H
+#define LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_POWERPC_H
 
-#include "Register/Arch/lldb-ppc64le-register-enums.h"
-#include "RegisterInfoInterface.h"
-#include "Utility/PPC64LE_DWARF_Registers.h"
+#include "Register/RegisterInfoInterface.h"
+#include "Register/Arch/RegisterContext_powerpc.h"
 #include "lldb/Target/RegisterContext.h"
 #include "lldb/Utility/Log.h"
 
-class RegisterContextPOSIX_ppc64le : public lldb_private::RegisterContext {
+class RegisterContextPOSIX_powerpc : public lldb_private::RegisterContext {
 public:
-  RegisterContextPOSIX_ppc64le(
+  RegisterContextPOSIX_powerpc(
       lldb_private::Thread &thread, uint32_t concrete_frame_idx,
       lldb_private::RegisterInfoInterface *register_info);
+
+  ~RegisterContextPOSIX_powerpc() override;
+
+  void Invalidate();
 
   void InvalidateAllRegisters() override;
 
@@ -40,19 +43,13 @@ public:
   const char *GetRegisterName(unsigned reg);
 
 protected:
-  // 64-bit general purpose registers.
-  uint64_t m_gpr_ppc64le[k_num_gpr_registers_ppc64le];
-
-  // floating-point registers including extended register.
-  uint64_t m_fpr_ppc64le[k_num_fpr_registers_ppc64le];
-
-  // VMX registers.
-  uint64_t m_vmx_ppc64le[k_num_vmx_registers_ppc64le * 2];
-
-  // VSX registers.
-  uint64_t m_vsx_ppc64le[k_num_vsx_registers_ppc64le * 2];
-
-  std::unique_ptr<lldb_private::RegisterInfoInterface> m_register_info_up;
+  uint64_t
+      m_gpr_powerpc[k_num_gpr_registers_powerpc]; // general purpose registers.
+  uint64_t
+      m_fpr_powerpc[k_num_fpr_registers_powerpc]; // floating point registers.
+  uint32_t m_vmx_powerpc[k_num_vmx_registers_powerpc][4];
+  std::unique_ptr<lldb_private::RegisterInfoInterface>
+      m_register_info_up; // Register Info Interface (FreeBSD or Linux)
 
   // Determines if an extended register set is supported on the processor
   // running the inferior process.
@@ -66,8 +63,12 @@ protected:
 
   bool IsVMX(unsigned reg);
 
-  bool IsVSX(unsigned reg);
-
+  virtual bool ReadGPR() = 0;
+  virtual bool ReadFPR() = 0;
+  virtual bool ReadVMX() = 0;
+  virtual bool WriteGPR() = 0;
+  virtual bool WriteFPR() = 0;
+  virtual bool WriteVMX() = 0;
 };
 
-#endif // LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_PPC64LE_H
+#endif // LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_POWERPC_H
