+++ lldb/source/Plugins/Process/Common/RegisterContext/POSIX/RegisterContextPOSIX_ppc64le.h
@@ -1,4 +1,4 @@
-//===-- RegisterContextPOSIX_s390x.h ----------------------------*- C++ -*-===//
+//===-- RegisterContextPOSIX_ppc64le.h --------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,28 +6,26 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_S390X_H
-#define LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_S390X_H
+#ifndef LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_PPC64LE_H
+#define LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_PPC64LE_H
 
-#include "RegisterContext_s390x.h"
-#include "RegisterInfoInterface.h"
-#include "lldb-s390x-register-enums.h"
+#include "Register/Arch/lldb-ppc64le-register-enums.h"
+#include "Register/RegisterInfoInterface.h"
+#include "Utility/PPC64LE_DWARF_Registers.h"
 #include "lldb/Target/RegisterContext.h"
 #include "lldb/Utility/Log.h"
 
-class RegisterContextPOSIX_s390x : public lldb_private::RegisterContext {
+class RegisterContextPOSIX_ppc64le : public lldb_private::RegisterContext {
 public:
-  RegisterContextPOSIX_s390x(
+  RegisterContextPOSIX_ppc64le(
       lldb_private::Thread &thread, uint32_t concrete_frame_idx,
       lldb_private::RegisterInfoInterface *register_info);
-
-  ~RegisterContextPOSIX_s390x() override;
-
-  void Invalidate();
 
   void InvalidateAllRegisters() override;
 
   size_t GetRegisterCount() override;
+
+  virtual size_t GetGPRSize();
 
   virtual unsigned GetRegisterSize(unsigned reg);
 
@@ -42,19 +40,22 @@ public:
   const char *GetRegisterName(unsigned reg);
 
 protected:
-  struct RegInfo {
-    uint32_t num_registers;
-    uint32_t num_gpr_registers;
-    uint32_t num_fpr_registers;
+  // 64-bit general purpose registers.
+  uint64_t m_gpr_ppc64le[k_num_gpr_registers_ppc64le];
 
-    uint32_t last_gpr;
-    uint32_t first_fpr;
-    uint32_t last_fpr;
-  };
+  // floating-point registers including extended register.
+  uint64_t m_fpr_ppc64le[k_num_fpr_registers_ppc64le];
 
-  RegInfo m_reg_info;
+  // VMX registers.
+  uint64_t m_vmx_ppc64le[k_num_vmx_registers_ppc64le * 2];
+
+  // VSX registers.
+  uint64_t m_vsx_ppc64le[k_num_vsx_registers_ppc64le * 2];
+
   std::unique_ptr<lldb_private::RegisterInfoInterface> m_register_info_up;
 
+  // Determines if an extended register set is supported on the processor
+  // running the inferior process.
   virtual bool IsRegisterSetAvailable(size_t set_index);
 
   virtual const lldb_private::RegisterInfo *GetRegisterInfo();
@@ -63,10 +64,10 @@ protected:
 
   bool IsFPR(unsigned reg);
 
-  virtual bool ReadGPR() = 0;
-  virtual bool ReadFPR() = 0;
-  virtual bool WriteGPR() = 0;
-  virtual bool WriteFPR() = 0;
+  bool IsVMX(unsigned reg);
+
+  bool IsVSX(unsigned reg);
+
 };
 
-#endif // LLDB_SOURCE_PLUGINS_PROCESS_UTILITY_REGISTERCONTEXTPOSIX_S390X_H
+#endif // LLDB_SOURCE_PLUGINS_PROCESS_COMMON_REGISTERCONTEXT_POSIX_REGISTERCONTEXTPOSIX_PPC64LE_H
