+++ llvm/utils/gn/secondary/lldb/source/Plugins/Process/Utility/BUILD.gn
@@ -41,15 +41,6 @@ static_library("Utility") {
     "RegisterContextMach_arm.cpp",
     "RegisterContextMach_x86_64.cpp",
     "RegisterContextMemory.cpp",
-    "RegisterContextPOSIX_arm.cpp",
-    "RegisterContextPOSIX_arm64.cpp",
-    "RegisterContextPOSIX_loongarch64.cpp",
-    "RegisterContextPOSIX_powerpc.cpp",
-    "RegisterContextPOSIX_ppc64le.cpp",
-    "RegisterContextPOSIX_riscv32.cpp",
-    "RegisterContextPOSIX_riscv64.cpp",
-    "RegisterContextPOSIX_s390x.cpp",
-    "RegisterContextPOSIX_x86.cpp",
     "RegisterContextThreadMemory.cpp",
     "StopInfoMachException.cpp",
     "ThreadMemory.cpp",
